backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
bcrypt==4.0.1
pydantic==2.4.2
psycopg2-binary
pymongo==4.5.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
Pillow==10.0.1
email-validator==2.0.0
python-json-logger==2.0.7
gunicorn==21.2.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0 